Moved AuthenticationService to `services` directory
