+++ src/app/sign-up/sign-up.page.ts
@@ -5,7 +5,7 @@ import {
   FormGroup,
   FormControl,
 } from '@angular/forms';
-import { AuthenticationService } from '../authentication.service';
+import { AuthenticationService } from '../services/authentication.service';
 import { Router } from '@angular/router';
 
 @Component({
